feat: fill*() binop modifiers (#135)
* feat: fill*() binop modifiers

* refactor: improve parser and avoid unwrap

* feat: allow fill/fill_left/fill_right as metric identifier

* chore: restore test being removed

* refactor: remove unneeded branch

+++ build.rs
@@ -17,10 +17,10 @@ use lrlex::{ct_token_map, DefaultLexerTypes};
 use lrpar::CTParserBuilder;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let ctp = CTParserBuilder::<DefaultLexerTypes<u8>>::new()
+    let ctp = CTParserBuilder::<DefaultLexerTypes<u16>>::new()
         .yacckind(YaccKind::Grmtools)
         .recoverer(lrpar::RecoveryKind::None)
         .grammar_in_src_dir("parser/promql.y")?
         .build()?;
-    ct_token_map::<u8>("token_map", ctp.token_map(), None)
+    ct_token_map::<u16>("token_map", ctp.token_map(), None)
 }
